Jokes

Documentation

Functions

getJoke

Arguments

:: IO (Either String Joke)
Joke response or error message

Read joke from web site. The Either type is used for error handling, so we can return error message if something went wrong.

tellJoke

Arguments

:: Joke
Joke

-> String
Formatted joke as setup and punchline

Print the setup and punchline of a joke.

Types

data Joke

A response from the joke web site.

Constructors

Joke

Fields

_type :: Text

Type of joke

_setup :: Text

Setup the joke

_punchline :: Text

Punchline

_id :: Int

Identifier